Lesson 24

+++ app/Http/Controllers/ThreadController.php
@@ -112,7 +112,7 @@ class ThreadController extends Controller
      */
     public function destroy($channel,Thread $thread)
     {
-        // $thread->replies()->delete();
+        $this->authorize('update',$thread);
         $thread->delete();
         if (request()->wantsJson()){
             return response([],204);
